backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlmodel==0.0.19
alembic==1.13.1
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9
pydantic-settings==2.2.1
resend==0.8.0
httpx==0.27.0
